seperate docx_manipulate.py
- rename email_related.py to email_utils.py
- divide docx_manipulate.py to docx_utils.py and pdf_utils.py
- update the README.mad
